+++ iModelCore/ECDb/ECDb/ProfileManager.cpp
@@ -142,6 +142,9 @@ DbResult ProfileManager::RunUpgraders(ProfileVersion const& versionBeforeUpgrade
     if (versionBeforeUpgrade < ProfileVersion(4, 0, 0, 3))
         upgraders.push_back(std::make_unique<ProfileUpgrader_4003>());
 
+    if (versionBeforeUpgrade < ProfileVersion(4, 0, 0, 4))
+        upgraders.push_back(std::make_unique<ProfileUpgrader_4004>());
+
     for (std::unique_ptr<ProfileUpgrader> const& upgrader : upgraders)
         {
         DbResult stat = upgrader->Upgrade(m_ecdb);